[rename] slideEvent.js의 이름을 slider.js로 변경
